Updated some expressions to use float

+++ include/timeMagic.cpp
@@ -9,11 +9,11 @@ void Timeline::init() {
 
 float Timeline::update() {
     Uint64 now = SDL_GetTicks();
-    double delta = (now - lastTicks) / 1000.0;
+    float delta = (now - lastTicks) / 1000.0f;
     lastTicks = now;
 
     if (isPaused()) {
-        return 0.0;
+        return 0.0f;
     } else {
         accumulated += delta * timeScale;
         return accumulated;
